requirements.txt
streamlit==1.47.0
tensorflow==2.19.0
pillow==11.3.0
numpy==2.1.3
inference_sdk